+++ server/database/migrations/2024_06_07_101817_images.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('images', function (Blueprint $table) {
             $table->id();
-            $table->integer('UsersID');
+            $table->unsignedBigInteger('UsersID');
             $table->string('ImageName');
             $table->timestamps();
             $table->foreign('UsersID')
